Update EveCentralPrices.gs
If the prices are added to a separate tab, it makes sense to allow the user to add headers for each column.

Personally I prefer using a combination of Index and Match functions instead of VLookup. I made an example googlesheet to show how that could work. The example also shows how the Filter function can provide a useful addition.

Lastly, I've changed the join from typeid to a comma separated list, for readability.

If a wrong typeID is submitted to the custom function, the whole thing grinds to a halt. Try looking up typeID 1861.
I've not figured out how to get around that as yet.